[release test] move cluster env utils to `anyscale_util`
as cluster envs are anyscale specific concepts
